refactor: extract TrackerIssueCandidate.swift from TrackerExportSupport.swift
Extracts the `TrackerIssueCandidate` data struct from
TrackerExportSupport.swift into its own file so TrackerExportSupport
holds only the shared helper enum (name matches file).

Byte-identical move of lines 3-8. Both files import Foundation.

TrackerExportSupport.swift: 206 → 199 lines. Tests: 667.
