+++ Sources/BugNarrator/Services/TrackerIssueCandidate.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct TrackerIssueCandidate: Equatable {
+    let remoteIdentifier: String
+    let title: String
+    let summary: String
+    let remoteURL: URL?
+}
